[AssemblyBrowser] Loading is now lock free
