+++ main/src/addins/MonoDevelop.AssemblyBrowser/MonoDevelop.AssemblyBrowser/TreeNodes/DomFieldNodeBuilder.cs
@@ -81,7 +81,7 @@ namespace MonoDevelop.AssemblyBrowser
 		{
 			if (DomMethodNodeBuilder.HandleSourceCodeEntity (navigator, data)) 
 				return null;
-			var field = CecilLoader.GetCecilObject ((IUnresolvedField)navigator.DataItem);
+			var field = GetCecilLoader (navigator).GetCecilObject ((IUnresolvedField)navigator.DataItem);
 			if (field == null)
 				return null;
 			return DomMethodNodeBuilder.Disassemble (data, rd => rd.DisassembleField (field));
@@ -91,7 +91,7 @@ namespace MonoDevelop.AssemblyBrowser
 		{
 			if (DomMethodNodeBuilder.HandleSourceCodeEntity (navigator, data)) 
 				return null;
-			var field = CecilLoader.GetCecilObject ((IUnresolvedField)navigator.DataItem);
+			var field = GetCecilLoader (navigator).GetCecilObject ((IUnresolvedField)navigator.DataItem);
 			if (field == null)
 				return null;
 			return DomMethodNodeBuilder.Decompile (data, DomMethodNodeBuilder.GetModule (navigator), field.DeclaringType, b => b.AddField (field));
@@ -101,7 +101,7 @@ namespace MonoDevelop.AssemblyBrowser
 		{
 			if (DomMethodNodeBuilder.HandleSourceCodeEntity (navigator, data)) 
 				return null;
-			var field = CecilLoader.GetCecilObject ((IUnresolvedField)navigator.DataItem);
+			var field = GetCecilLoader (navigator).GetCecilObject ((IUnresolvedField)navigator.DataItem);
 			if (field == null)
 				return null;
 			return DomMethodNodeBuilder.GetSummary (data, DomMethodNodeBuilder.GetModule (navigator), field.DeclaringType, b => b.AddField (field));
